src/backend/requirements.txt
apscheduler==3.11.0
asgiref==3.8.1
black==25.1.0
confluent-kafka==2.10.0
coverage==7.8.0
Django==4.2.20
django-apscheduler==0.7.0
django-cors-headers==4.0.0
django-stubs==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==23.0.0
isort==6.0.1
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.8.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.10.0
pytest-mock==3.14.0
pytz==2025.1
ruff==0.9.10
setuptools==76.0.0
sqlparse==0.5.3
